SaFliTe/requirements.txt
openai==1.0.0
llama-cpp-python